+++ datanode/networkhistory/config.go
@@ -19,6 +19,9 @@ type Config struct {
 	Store    store.Config    `group:"Store" namespace:"store"`
 	Snapshot snapshot.Config `group:"Snapshot" namespace:"snapshot"`
 
+	FetchRetryMax int               `long:"fetch-retry-max" description:"maximum number of times to retry fetching segments - default 10"`
+	RetryTimeout  encoding.Duration `long:"retry-timeout" description:"time to wait between retries, increases with each retry - default 5s"`
+
 	Initialise InitializationConfig `group:"Initialise" namespace:"initialise"`
 }
 
@@ -32,6 +35,8 @@ func NewDefaultConfig() Config {
 		Publish:       true,
 		Store:         store.NewDefaultConfig(),
 		Snapshot:      snapshot.NewDefaultConfig(),
+		FetchRetryMax: 10,
+		RetryTimeout:  encoding.Duration{Duration: 5 * time.Second},
 		Initialise:    NewDefaultInitializationConfig(),
 	}
 }
@@ -42,8 +47,6 @@ func NewDefaultInitializationConfig() InitializationConfig {
 		TimeOut:           encoding.Duration{Duration: 1 * time.Minute},
 		GrpcAPIPorts:      []int{},
 		ToSegment:         "",
-		FetchRetryMax:     5,
-		RetryTimeout:      encoding.Duration{Duration: time.Second},
 	}
 }
 
@@ -52,6 +55,4 @@ type InitializationConfig struct {
 	MinimumBlockCount int64             `long:"block-count" description:"the minimum number of blocks to fetch"`
 	TimeOut           encoding.Duration `long:"timeout" description:"maximum time allowed to auto-initialise the node"`
 	GrpcAPIPorts      []int             `long:"grpc-api-ports" description:"list of additional ports to check to for api connection when getting latest segment"`
-	FetchRetryMax     uint64            `long:"fetch-retry-max" description:"maximum number of times to retry fetching segments - default 5"`
-	RetryTimeout      encoding.Duration `long:"retry-timeout" description:"time to wait between retries - default 1 second"`
 }